+++ src/app/chatbot-design-studio/cds-dashboard/cds-canvas/actions/list-voice/elements/cds-action-reply-audio/cds-action-reply-audio.component.ts
@@ -53,7 +53,7 @@ export class CdsActionReplyAudioComponent implements OnInit {
 
 
   private initialize(){
-    this.delayTime = this.wait.time? (this.wait.time/1000) : 500;
+    this.delayTime = (this.wait && this.wait.time  || this.wait.time === 0)? (this.wait.time/1000) : 500/1000;
     if(this.response && this.response._tdJSONCondition && this.response._tdJSONCondition.conditions.length > 0){
       this.filterConditionExist = true
     }