+++ src/app/chatbot-design-studio/cds-dashboard/cds-canvas/actions/list-voice/elements/cds-action-reply-voice-text/cds-action-reply-voice-text.component.ts
@@ -91,7 +91,7 @@ export class CdsActionReplyVoiceTextComponent implements OnInit {
   // PRIVATE FUNCTIONS //
 
   private initialize(){
-    this.delayTime = (this.wait && this.wait.time)? (this.wait.time/1000) : 500;
+    this.delayTime = (this.wait && this.wait.time  || this.wait.time === 0)? (this.wait.time/1000) : 500/1000;
     this.checkButtons();
     this.buttons = this.intentService.patchButtons(this.buttons, this.idAction);
     this.idIntent = this.idAction.split('/')[0];
@@ -186,6 +186,7 @@ export class CdsActionReplyVoiceTextComponent implements OnInit {
     this.delayTime = value;
     this.wait.time = value*1000;
     this.canShowFilter = true;
+    console.log('delayyyyyyy', this.delayTime)
     this.changeActionReply.emit();
   }