+++ src/app/chatbot-design-studio/cds-dashboard/cds-canvas/actions/list/cds-action-reply/elements/cds-action-reply-image/cds-action-reply-image.component.ts
@@ -81,7 +81,7 @@ export class CdsActionReplyImageComponent implements OnInit {
   // PRIVATE FUNCTIONS //
 
   private initialize(){
-    this.delayTime = (this.wait && this.wait.time)? (this.wait.time/1000) : 500;
+    this.delayTime = (this.wait && this.wait.time  || this.wait.time === 0)? (this.wait.time/1000) : 500/1000;
     this.checkButtons();
     // this.patchButtons();
     // this.buttons = this.response?.attributes?.attachment?.buttons;