+++ src/app/chatbot-design-studio/cds-dashboard/cds-canvas/actions/list/cds-action-reply/elements/cds-action-reply-text/cds-action-reply-text.component.ts
@@ -79,7 +79,7 @@ export class CdsActionReplyTextComponent implements OnInit {
   // PRIVATE FUNCTIONS //
 
   private initialize(){
-    this.delayTime = (this.wait && this.wait.time)? (this.wait.time/1000) : 500;
+    this.delayTime = (this.wait && this.wait.time  || this.wait.time === 0)? (this.wait.time/1000) : 500/1000;
     this.checkButtons();
     this.buttons = this.intentService.patchButtons(this.buttons, this.idAction);
     this.idIntent = this.idAction.split('/')[0];