+++ src/app/chatbot-design-studio/services/intent.service.ts
@@ -899,6 +899,7 @@ export class IntentService {
       action.top_k = 5;
       action.model = TYPE_GPT_MODEL['GPT-4'].value
       action.preview = [];
+      action.history = false;
     }
     if(typeAction === TYPE_ACTION.GPT_TASK){
       action = new ActionGPTTask();
@@ -1017,7 +1018,7 @@ export class IntentService {
     }
     if(typeAction === TYPE_ACTION_VXML.SPEECH_FORM){
       action = new ActionVoice(TYPE_ACTION_VXML.SPEECH_FORM);
-      (action as ActionVoice).attributes.disableInputMessage = false
+      // (action as ActionVoice).attributes.disableInputMessage = false
       let commandWait = new Wait();
       action.attributes.commands.push(commandWait);
       let command = new Command(TYPE_COMMAND.MESSAGE);