+++ CHANGELOG.md
@@ -6,6 +6,21 @@
     *Giovanni Troisi*  
 ### **Copyrigth**: *Tiledesk SRL*
 
+### 1.12.0-rc.5
+👉 **changed**: namescace model 
+👉 **changed**: top_k range in action askkb_v2
+
+### 1.12.0-rc.4
+### 1.12.0-rc.3
+👉 **added**: namespace, context and history properties into action-askgptv2
+
+### 1.12.0-rc.2
+👉 **removed**: disableInput option from voice actions
+
+### 1.12.0-rc.1
+👉 **changed**: voiceFlow variables restored
+👉 **bug-fixed**: wait time in action-reply and action-reply-voice is incorrect if set to 0
+
 ### 1.11.0 in PROD
 
 ### 1.11.0-rc.1